fix: resolve web startup budget failure
